requirements.txt
streamlit
pandas
numpy
torch
altair
requests
openpyxl
xlrd
pyyaml